+++ src/test/java/us/kbase/test/auth2/lib/identity/RemoteIdentityTest.java
@@ -21,10 +21,10 @@ public class RemoteIdentityTest {
 		assertThat("incorrect username", dets.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets.getFullname(), is("full"));
 		assertThat("incorrect email", dets.getEmail(), is("email"));
-		assertThat("incorrect mfa authenticated", dets.getMfaAuthenticated(), is(MfaStatus.UNKNOWN));
+		assertThat("incorrect mfa authenticated", dets.getMfa(), is(MfaStatus.UNKNOWN));
 		assertThat("incorrect hashcode", dets.hashCode(), is(-497844993));
 		assertThat("incorrect toString()", dets.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfaAuthenticated=UNKNOWN]"));
+				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfa=UNKNOWN]"));
 	}
 	
 	@Test
@@ -33,19 +33,19 @@ public class RemoteIdentityTest {
 		assertThat("incorrect username", dets.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets.getFullname(), is((String) null));
 		assertThat("incorrect email", dets.getEmail(), is((String) null));
-		assertThat("incorrect mfa authenticated", dets.getMfaAuthenticated(), is(MfaStatus.UNKNOWN));
+		assertThat("incorrect mfa authenticated", dets.getMfa(), is(MfaStatus.UNKNOWN));
 		assertThat("incorrect hashcode", dets.hashCode(), is(4522828));
 		assertThat("incorrect toString()", dets.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=null, email=null, mfaAuthenticated=UNKNOWN]"));
+				is("RemoteIdentityDetails [username=user, fullname=null, email=null, mfa=UNKNOWN]"));
 		
 		final RemoteIdentityDetails dets2 = new RemoteIdentityDetails("user", null, null);
 		assertThat("incorrect username", dets2.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets2.getFullname(), is((String) null));
 		assertThat("incorrect email", dets2.getEmail(), is((String) null));
-		assertThat("incorrect mfa authenticated", dets2.getMfaAuthenticated(), is(MfaStatus.UNKNOWN));
+		assertThat("incorrect mfa authenticated", dets2.getMfa(), is(MfaStatus.UNKNOWN));
 		assertThat("incorrect hashcode", dets2.hashCode(), is(4522828));
 		assertThat("incorrect toString()", dets2.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=null, email=null, mfaAuthenticated=UNKNOWN]"));
+				is("RemoteIdentityDetails [username=user, fullname=null, email=null, mfa=UNKNOWN]"));
 	}
 	
 	@Test
@@ -119,7 +119,7 @@ public class RemoteIdentityTest {
 		assertThat("incorrect hashcode", ri.hashCode(), is(124952370));
 		assertThat("incorrect toString()", ri.toString(),
 				is("RemoteIdentity [remoteID=RemoteIdentityID [provider=p, id=i], " +
-						"details=RemoteIdentityDetails [username=u, fullname=f, email=e, mfaAuthenticated=UNKNOWN]]"));
+						"details=RemoteIdentityDetails [username=u, fullname=f, email=e, mfa=UNKNOWN]]"));
 	}
 	
 	@Test
@@ -217,9 +217,9 @@ public class RemoteIdentityTest {
 		assertThat("incorrect username", dets.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets.getFullname(), is("full"));
 		assertThat("incorrect email", dets.getEmail(), is("email"));
-		assertThat("incorrect mfa authenticated", dets.getMfaAuthenticated(), is(MfaStatus.USED));
+		assertThat("incorrect mfa authenticated", dets.getMfa(), is(MfaStatus.USED));
 		assertThat("incorrect toString()", dets.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfaAuthenticated=USED]"));
+				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfa=USED]"));
 	}
 
 	@Test
@@ -228,9 +228,9 @@ public class RemoteIdentityTest {
 		assertThat("incorrect username", dets.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets.getFullname(), is("full"));
 		assertThat("incorrect email", dets.getEmail(), is("email"));
-		assertThat("incorrect mfa authenticated", dets.getMfaAuthenticated(), is(MfaStatus.NOT_USED));
+		assertThat("incorrect mfa authenticated", dets.getMfa(), is(MfaStatus.NOT_USED));
 		assertThat("incorrect toString()", dets.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfaAuthenticated=NOT_USED]"));
+				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfa=NOT_USED]"));
 	}
 
 	@Test
@@ -239,9 +239,9 @@ public class RemoteIdentityTest {
 		assertThat("incorrect username", dets.getUsername(), is("user"));
 		assertThat("incorrect fullname", dets.getFullname(), is("full"));
 		assertThat("incorrect email", dets.getEmail(), is("email"));
-		assertThat("incorrect mfa authenticated", dets.getMfaAuthenticated(), is(MfaStatus.UNKNOWN));
+		assertThat("incorrect mfa authenticated", dets.getMfa(), is(MfaStatus.UNKNOWN));
 		assertThat("incorrect toString()", dets.toString(),
-				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfaAuthenticated=UNKNOWN]"));
+				is("RemoteIdentityDetails [username=user, fullname=full, email=email, mfa=UNKNOWN]"));
 	}
 
 	@Test