+++ src/test/java/us/kbase/test/auth2/service/api/APITokenTest.java
@@ -40,7 +40,7 @@ public class APITokenTest {
 		
 		assertThat("incorrect id", t.getId(), is(id.toString()));
 		assertThat("incorrect cache time", t.getCachefor(), is(20000L));
-		assertThat("incorrect mfa status", t.getMfaAuthenticated(), is(MfaStatus.UNKNOWN));
+		assertThat("incorrect mfa status", t.getMfa(), is(MfaStatus.UNKNOWN));
 	}